project/pyproject.toml
[project]
name = "project"
version = "0.0.1"
authors = [{ name = "magnus@mecoding.se" }]
requires-python = ">=3.10,<=3.14"
dependencies = [
    # Any dependencies that should be included in the deploy to databricks can be added here
]
[dependency-groups]
dev = [
    "delta-spark==4.0.1",
    "pyspark==4.1.1",
    "databricks-sdk==0.85.0",
    "pytest",
    "ruff",
    "databricks-dlt",
    "databricks-connect>=15.4,<15.5",
    "ipykernel",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120

 [tool.hatch.build.targets.wheel]
      packages = ["src"]

